Move Time class to a separate file time.py
